# Fix performance issue in apisChangeLifecyclePost event (Fix https://github.com/wso2/product-apim/issues/3910)