--- a/components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/APIProviderImpl.java
+++ b/components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/APIProviderImpl.java
@@ -1451,8 +1451,7 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
                             failedGateways = publishToGateway(api);
                             //Sending Notifications to existing subscribers
                             if (APIConstants.PUBLISHED.equals(newStatus)) {
-                                List<APIIdentifier> oldPublishedAPIList = getOldPublishedAPIList(api);
-                                sendEmailNotification(api, oldPublishedAPIList);
+                                sendEmailNotification(api);
                             }
                         } else { // API Status : RETIRED or CREATED
                             failedGateways = removeFromGateway(api);
@@ -1691,10 +1690,9 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
      * This method used to send notifications to the previous subscribers of older versions of a given API
      *
      * @param api
-     * @param apiIdentifiers
      * @throws APIManagementException
      */
-    private void sendEmailNotification(API api, List<APIIdentifier> apiIdentifiers) throws APIManagementException {
+    private void sendEmailNotification(API api) throws APIManagementException {
         try {
             String isNotificationEnabled = "false";
             Registry configRegistry = ServiceReferenceHolder.getInstance().getRegistryService().
@@ -1708,6 +1706,7 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
                 }
             }
             if (JavaUtils.isTrueExplicitly(isNotificationEnabled)) {
+                List<APIIdentifier> apiIdentifiers = getOldPublishedAPIList(api);
                 for (APIIdentifier oldAPI : apiIdentifiers) {
                     Properties prop = new Properties();
                     prop.put(NotifierConstants.API_KEY, oldAPI);