## Purpose
org.wso2.carbon.apimgt.impl.APIProviderImpl.getAPIsByProvider() method consumes a significant time (903ms) when compared to the total response time 2088ms.

Local:
[2018-10-26 16:33:27,093] INFO - APIProviderImpl Execution time:: getOldPublishedAPIList : getAPIsByProvider 903ms
[2018-10-26 16:33:26,188] INFO - APIProviderImpl Execution time:: propergateAPIStatusChangeToGateways : publishToGateway 739ms

When analyzing, getAPIsByProvider method invocation happens even when the Send Email notification feature deactivated.

## Goals
Reduce the time taken to publish APIs when there are large numbers of APIs present in Publisher.

## Approach
As the solution, getAPIsByProvider method was moved inside the isNotificationEnabled logic so that this method won't invoke unless the Send Email notification feature has activated.

## User stories
Publish an API

## Release note
Bug Fix

## Documentation
No Doc Impact

## Training
N/A

## Certification
N/A

## Marketing
N/A

## Automation tests

## Security checks
 - Followed secure coding standards in http://wso2.com/technical-reports/wso2-secure-engineering-guidelines? yes
 - Ran FindSecurityBugs plugin and verified report? yes/no
 - Confirmed that this PR doesn't commit any keys, passwords, tokens, usernames, or other secrets? yes

## Samples
N/A

## Related PRs
N/A

## Migrations (if applicable)
N/A
## Learning
N/A